uint8 to uint32 conversion

+++ include/block.h
@@ -24,7 +24,7 @@ struct Block : GameObject {
     Block();
 	~Block() override;
 	void update(GameObject** game_objects_p,
-		    uint8 game_objects_size) override;
+		    	uint32 game_objects_size) override;
 };
 
 #endif
